+++ hsml.py
@@ -18,14 +18,14 @@ def get_smooth_length(bar):
     """
     #Are we arepo? If we are a modern version we should have this array.
     if np.any(np.array(bar.keys()) == 'Volume'):
-        volume=np.array(bar["Volume"],dtype=np.float64)
+        volume=np.array(bar["Volume"])
         radius = (3*volume/4/math.pi)**(0.33333333)
     elif np.any(np.array(bar.keys()) == 'Number of faces of cell'):
-        rho=np.array(bar["Density"],dtype=np.float64)
-        mass=np.array(bar["Masses"],dtype=np.float64)
+        rho=np.array(bar["Density"])
+        mass=np.array(bar["Masses"])
         volume = mass/rho
         radius = (3*volume/4/math.pi)**(0.33333333)
     else:
         #If we are gadget, the SmoothingLength array is actually the smoothing length.
-        radius=np.array(bar["SmoothingLength"],dtype=np.float64)
+        radius=np.array(bar["SmoothingLength"])
     return radius
